+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -37,16 +37,4 @@ public class ReservationController {
         reservationService.deleteById(id, member);
         return ResponseEntity.noContent().build();
     }
-
-    @ExceptionHandler({NotExistEntityException.class,
-            DuplicateEntityException.class,
-            InvalidAuthorizationTokenException.class})
-    public ResponseEntity handleBadRequest() {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(NotQualifiedMemberException.class)
-    public ResponseEntity handleUnauthorized() {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
 }
